not user layout fix

+++ app/controllers/pages_controller.rb
@@ -22,28 +22,32 @@ class PagesController < ApplicationController
   def index
     @novel = Novel.find(params[:novel_id])
     @pages = @novel.pages
-    @favorite = Favorite.find_by(user_id:current_user.id, novel_id: @novel)
+    if user_signed_in?
+      @favorite = Favorite.find_by(user_id:current_user.id, novel_id: @novel)
+    end
 
   end
 
   def show
     @page = Page.find(params[:id])
     @novel = Novel.find(@page.novel_id)
-    new_history = @novel.histories.new
-    new_history.user_id = current_user.id
-    new_history.page_id = @page.id
+    if user_signed_in?
+      new_history = @novel.histories.new
+      new_history.user_id = current_user.id
+      new_history.page_id = @page.id
 
-    if current_user.histories.exists?(novel_id: "#{params[:novel_id]}")
-      old_history = current_user.histories.find_by(novel_id: "#{params[:novel_id]}")
-      old_history.destroy
-    end
+      if current_user.histories.exists?(novel_id: "#{params[:novel_id]}")
+        old_history = current_user.histories.find_by(novel_id: "#{params[:novel_id]}")
+        old_history.destroy
+      end
 
-    new_history.save
+      new_history.save
 
-    histories_stock_limit = 10
-    histories = current_user.histories.all
-    if histories.count > histories_stock_limit
-      histories[0].destroy
+      histories_stock_limit = 10
+      histories = current_user.histories.all
+      if histories.count > histories_stock_limit
+        histories[0].destroy
+      end
     end
 
     @novel.pages.count
